requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==5.2.1
deprecated==1.2.18
google==3.0.0
googleapis-common-protos==1.70.0
idna==3.10
importlib-metadata==8.6.1
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
protobuf==5.29.5
py==1.11.0
pycryptodome==3.9.9
pydantic==2.11.5
pydantic-core==2.33.2
python-dotenv==1.1.0
pytz==2020.5
requests==2.32.3
retry==0.9.2
six==1.17.0
soupsieve==2.7
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
volcengine==1.0.187
wrapt==1.17.2
zipp==3.22.0
